Introduce the service plugin type
