+++ libraries/src/Plugin/CMSServicePlugin.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Joomla! Content Management System
+ *
+ * @copyright  Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+namespace Joomla\CMS\Plugin;
+
+defined('JPATH_PLATFORM') or die;
+
+/**
+ * Plugin Class
+ *
+ * @since  __DEPLOY_VERSION__
+ */
+abstract class CMSServicePlugin
+{
+}
